requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic-settings
email-validator
alembic
